refactor: lifecycle add app
